# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RemoteSupportLiveKit backend",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create Session",
        "description": "Create a session. Auth: none. Rate limit: 5/min/IP, 20/hour/device.",
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions/{session_id}/agent": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Set Ai Enabled",
        "description": "Enable/disable the AI agent's speech. **Support auth only** (FR-4.1).",
        "operationId": "set_ai_enabled_v1_sessions__session_id__agent_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/agent/mode": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Set Agent Mode",
        "description": "Record a mode transition. Called by the agent worker (service auth) or\nforced by support.",
        "operationId": "set_agent_mode_v1_sessions__session_id__agent_mode_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions/{session_id}/recordings": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "List Recordings",
        "description": "List the session's media with V4 signed URLs.\n\nThe backend never proxies media bytes -- egress writes straight to GCS and\nclients fetch via short-lived signed URLs (SIGNED_URL_TTL_SECONDS).",
        "operationId": "list_recordings_v1_sessions__session_id__recordings_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentMode": {
        "type": "string",
        "enum": [
          "SOLO",
          "ASSISTED",
          "WRAP_UP"
        ],
        "title": "AgentMode"
      },
      "AgentModeRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AgentMode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "AgentModeRequest"
      },
      "AgentToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "AgentToggleRequest"
      },
      "AgentToggleResponse": {
        "properties": {
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "metadata_version": {
            "type": "integer",
            "title": "Metadata Version"
          },
          "applied_at": {
            "type": "string",
            "format": "date-time",
            "title": "Applied At"
          }
        },
        "type": "object",
        "required": [
          "ai_enabled",
          "metadata_version",
          "applied_at"
        ],
        "title": "AgentToggleResponse"
      },
...
      "CreateSessionRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en-US"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "CreateSessionRequest"
      },
      "CreateSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "join_code": {
            "type": "string",
            "title": "Join Code"
          },
          "join_code_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Join Code Expires At"
          },
          "deep_link": {
            "type": "string",
            "title": "Deep Link"
          },
          "universal_link": {
            "type": "string",
            "title": "Universal Link"
          },
          "qr_payload": {
            "type": "string",
            "title": "Qr Payload"
          },
          "caller_session_token": {
            "type": "string",
            "title": "Caller Session Token"
          },
          "consent_required": {
            "type": "boolean",
            "title": "Consent Required",
            "default": true
          },
          "consent_text_version": {
            "type": "string",
            "title": "Consent Text Version"
          },
          "consent_text": {
            "type": "string",
            "title": "Consent Text"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "join_code",
          "join_code_expires_at",
          "deep_link",
          "universal_link",
          "qr_payload",
          "caller_session_token",
          "consent_text_version",
          "consent_text"
        ],
        "title": "CreateSessionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParticipantRole": {
        "type": "string",
        "enum": [
          "caller",
          "support",
          "agent"
        ],
        "title": "ParticipantRole"
      },
      "RecordingInfo": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/RecordingKind"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParticipantRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/RecordingState"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "gcs_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gcs Uri"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "url_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Expires At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "state"
        ],
        "title": "RecordingInfo"
      },
      "RecordingKind": {
        "type": "string",
        "enum": [
          "track_video",
          "track_audio",
          "room_composite"
        ],
        "title": "RecordingKind"
      },
      "RecordingState": {
        "type": "string",
        "enum": [
          "starting",
          "active",
          "ending",
          "complete",
          "failed",
          "aborted"
        ],
        "title": "RecordingState"
      },
      "RecordingsResponse": {
        "properties": {
          "recordings": {
            "items": {
              "$ref": "#/components/schemas/RecordingInfo"
            },
            "type": "array",
            "title": "Recordings"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptExportUrls"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "recordings"
        ],
        "title": "RecordingsResponse"
      },
...
      "TranscriptExportUrls": {
        "properties": {
          "jsonl_url": {
            "type": "string",
            "title": "Jsonl Url"
          },
          "vtt_url": {
            "type": "string",
            "title": "Vtt Url"
          },
          "txt_url": {
            "type": "string",
            "title": "Txt Url"
          }
        },
        "type": "object",
        "required": [
          "jsonl_url",
          "vtt_url",
          "txt_url"
        ],
        "title": "TranscriptExportUrls"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}